fetching token list.
