+++ src/hooks/useTheme.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ThemeContext } from '@contexts';
+import { ThemeContext } from '../contexts/themeContext/ThemeContext';
 
 const useTheme = () => useContext(ThemeContext);
